refactor(cssg-plugin-assets): split code
