+++ packages/contentful-ssg/src/__test__/fixtures/assets.json
@@ -154,23 +154,23 @@
         "sys": {
           "type": "Link",
           "linkType": "Space",
-          "id": "focmz63cn6j9"
+          "id": "gpdredy5px7h"
         }
       },
-      "id": "3NIRCubNZhxuGRRca3zQMo",
+      "id": "1J3uqsCGfgXe8pWnEf55Iz",
       "type": "Asset",
-      "createdAt": "2022-05-25T13:58:24.835Z",
-      "updatedAt": "2022-05-25T13:58:37.038Z",
+      "createdAt": "2022-06-06T22:06:40.586Z",
+      "updatedAt": "2022-06-06T22:06:57.071Z",
       "environment": {
         "sys": {
-          "id": "develop",
+          "id": "master",
           "type": "Link",
           "linkType": "Environment"
         }
       },
-      "publishedVersion": 4,
-      "publishedAt": "2022-05-25T13:58:37.038Z",
-      "firstPublishedAt": "2022-05-25T13:58:37.038Z",
+      "publishedVersion": 3,
+      "publishedAt": "2022-06-06T22:06:57.071Z",
+      "firstPublishedAt": "2022-06-06T22:06:57.071Z",
       "createdBy": {
         "sys": {
           "type": "Link",
@@ -186,7 +186,7 @@
         }
       },
       "publishedCounter": 1,
-      "version": 5,
+      "version": 4,
       "publishedBy": {
         "sys": {
           "type": "Link",
@@ -197,19 +197,82 @@
     },
     "fields": {
       "title": {
-        "en-US": "video"
-      },
-      "description": {
-        "en-US": ""
+        "en-US": "file example MP4 640 3MG"
       },
       "file": {
         "en-US": {
-          "url": "//videos.ctfassets.net/focmz63cn6j9/3NIRCubNZhxuGRRca3zQMo/9846ea45648f018477b31b3647ec16ce/city-jimmy.mp4",
+          "url": "//videos.ctfassets.net/gpdredy5px7h/1J3uqsCGfgXe8pWnEf55Iz/993750eabc17cb4a568d3993d1c97ef9/file_example_MP4_640_3MG.mp4",
           "details": {
-            "size": 1512243
+            "size": 3114374
           },
-          "fileName": "city-jimmy.mp4",
+          "fileName": "file_example_MP4_640_3MG.mp4",
           "contentType": "video/mp4"
+        }
+      }
+    }
+  },
+  {
+    "metadata": {
+      "tags": []
+    },
+    "sys": {
+      "space": {
+        "sys": {
+          "type": "Link",
+          "linkType": "Space",
+          "id": "gpdredy5px7h"
+        }
+      },
+      "id": "5B0NQTKpQVNyPCwF7Ffd3e",
+      "type": "Asset",
+      "createdAt": "2022-06-06T21:46:37.849Z",
+      "updatedAt": "2022-06-06T21:46:48.772Z",
+      "environment": {
+        "sys": {
+          "id": "master",
+          "type": "Link",
+          "linkType": "Environment"
+        }
+      },
+      "publishedVersion": 3,
+      "publishedAt": "2022-06-06T21:46:48.772Z",
+      "firstPublishedAt": "2022-06-06T21:46:48.772Z",
+      "createdBy": {
+        "sys": {
+          "type": "Link",
+          "linkType": "User",
+          "id": "2lyDFdgxALbzi4FDNHc9x9"
+        }
+      },
+      "updatedBy": {
+        "sys": {
+          "type": "Link",
+          "linkType": "User",
+          "id": "2lyDFdgxALbzi4FDNHc9x9"
+        }
+      },
+      "publishedCounter": 1,
+      "version": 4,
+      "publishedBy": {
+        "sys": {
+          "type": "Link",
+          "linkType": "User",
+          "id": "2lyDFdgxALbzi4FDNHc9x9"
+        }
+      }
+    },
+    "fields": {
+      "title": {
+        "en-US": "Test"
+      },
+      "file": {
+        "en-US": {
+          "url": "//assets.ctfassets.net/gpdredy5px7h/5B0NQTKpQVNyPCwF7Ffd3e/4c30c592475882e0f3c76265f19db101/Test.pdf",
+          "details": {
+            "size": 599538
+          },
+          "fileName": "Test.pdf",
+          "contentType": "application/pdf"
         }
       }
     }
