+++ packages/cssg-plugin-assets/package.json
@@ -4,12 +4,13 @@
   "description": "",
   "type": "module",
   "main": "./dist/index.js",
-  "typings": "./dist/index.d.ts",
+  "typings": "./dist/types.d.ts",
   "exports": {
     ".": "./dist/index.js"
   },
   "devDependencies": {
     "@jungvonmatt/contentful-ssg": "^1.8.0-alpha.2",
+    "@types/svgo": "^2.6.3",
     "eslint": "^8.0.1",
     "eslint-config-xo": "^0.39.0",
     "eslint-config-xo-space": "^0.30.0",
