+++ packages/cssg-plugin-assets/src/index.ts
@@ -1,123 +1,16 @@
-/* eslint-disable complexity */
-import { createFFmpeg, fetchFile } from '@ffmpeg/ffmpeg';
-import { Asset, MapAssetLink, TransformContext } from '@jungvonmatt/contentful-ssg';
-import { Presets, SingleBar } from 'cli-progress';
-import { existsSync, promises } from 'fs';
-import got from 'got';
-import mkdirp from 'mkdirp';
-import { basename, dirname, extname as pathExtname, join, resolve } from 'path';
-import { optimize } from 'svgo';
-
-// Max width that can be handled by the contentful image api
-const contentfulMaxWidth = 4000;
-// Some image formats (avif) have an additional limitation on the megapixel size
-const maxMegaPixels = {
-  avif: 9,
-};
-
-export type Ratios = Record<string, number>;
-export type FocusArea =
-  | 'center'
-  | 'top'
-  | 'right'
-  | 'left'
-  | 'bottom'
-  | 'top_right'
-  | 'top_left'
-  | 'bottom_right'
-  | 'bottom_left'
-  | 'face'
-  | 'faces';
-
-export type RatioConfig = {
-  default?: Ratios;
-  contentTypes?: {
-    [contentTypeId: string]: {
-      default?: Ratios;
-      fields?: {
-        [fieldId: string]: Ratios;
-      };
-    };
-  };
-};
-
-export type FocusAreaConfig = {
-  default?: FocusArea;
-  contentTypes?: {
-    [contentTypeId: string]: {
-      default?: FocusArea;
-      fields?: {
-        [fieldId: string]: FocusArea;
-      };
-    };
-  };
-};
-
-// eslint-disable-next-line no-unused-vars
-export type SizesCallback = (asset: Asset, ratio: number, focusArea: string) => number;
-
-export interface PluginConfig {
-  sizes?: Array<number | SizesCallback>;
-  rootDir?: string;
-  assetBase?: string;
-  assetFolder?: string;
-  cacheFolder?: string;
-  cachePath?: string;
-  assetPath?: string;
-  extraTypes?: string[];
-  quality?: number;
-  ratios?: RatioConfig;
-  focusAreas?: FocusAreaConfig;
-  download?: boolean;
-  generatePosterImages?: boolean;
-  posterPosition?: string;
-  posterScale?: string;
-}
-
-export type ProcessedAsset = MapAssetLink & {
-  src: string;
-};
-
-export type ProcessedSvg = ProcessedAsset & {
-  source: string;
-  srcsets: [];
-};
-
-type Srcset = {
-  type: string;
-  srcset: string;
-};
-
-export type Derivative = {
-  width: number;
-  height: number;
-  sizes: string;
-  srcsets: Srcset[];
-  src: string;
-};
-
-export type ProcessedImage = ProcessedAsset & {
-  derivatives: {
-    original: Derivative;
-    [ratio: string]: Derivative;
-  };
-};
-
-export type ProcessedVideo = ProcessedAsset & {
-  poster?: string;
-};
-
-const getAssetTimestamp = async (sys: Asset['sys']) =>
-  Date.parse(sys?.updatedAt || sys?.createdAt) || 0;
-
-const getFileTimestamp = async (file) => {
-  if (existsSync(file)) {
-    const { mtime } = await promises.stat(file);
-    return Date.parse(mtime.toISOString());
-  }
-
-  return 0;
-};
+import type { MapAssetLink, TransformContext } from '@jungvonmatt/contentful-ssg';
+import { resolve } from 'path';
+import { getAssetHelper } from './helper/asset.js';
+import { getImageHelper } from './helper/image.js';
+import { getSvgHelper } from './helper/svg.js';
+import { getVideoHelper } from './helper/video.js';
+import type {
+  PluginConfig,
+  ProcessedAsset,
+  ProcessedImage,
+  ProcessedSvg,
+  ProcessedVideo,
+} from './types.js';
 
 /**
  *
@@ -125,7 +18,7 @@ const getFileTimestamp = async (file) => {
  * @returns
  */
 const processOptions = (options: PluginConfig = {}): PluginConfig => {
-  const defaultOptions = {
+  const defaultOptions: PluginConfig = {
     sizes: [1920, 1280, 640, 320],
     rootDir: process.cwd(),
     assetBase: '/assets/cf',
@@ -135,6 +28,17 @@ const processOptions = (options: PluginConfig = {}): PluginConfig => {
     quality: 80,
     ratios: {},
     focusAreas: {},
+    svgoPlugins: [
+      {
+        name: 'preset-default',
+        params: {
+          overrides: {
+            removeViewBox: false,
+          },
+        },
+      },
+      'removeDimensions',
+    ],
   };
 
   const { rootDir, cacheFolder, assetFolder } = { ...defaultOptions, ...options };
@@ -151,427 +55,47 @@ const processOptions = (options: PluginConfig = {}): PluginConfig => {
 };
 
 export default (pluginOptions?: PluginConfig) => {
-  const queue = new Set<string>();
-  const posterQueue = new Set<string>();
   const options = processOptions(pluginOptions);
 
-  const ffmpeg = createFFmpeg();
-
-  const getPosterImageFilePath = (videoFile) => `${videoFile.replace(/\.\w+$/, '')}-poster.jpg`;
-
-  const generatePosterImage = async (src, dest) => {
-    const srcPath = join(options.cachePath, src);
-    const cachedDestPath = join(options.cachePath, dest);
-    const destPath = join(options.assetPath, dest);
-
-    const timestampSrc = await getFileTimestamp(srcPath);
-    const timestampDestCached = await getFileTimestamp(cachedDestPath);
-    const timestampDest = await getFileTimestamp(cachedDestPath);
-
-    if (!timestampDestCached || timestampDestCached < timestampSrc) {
-      const ffmpegSrc = src.replace(/^\//, '').replace(/\//g, '-');
-      const ffmpegDest = dest.replace(/^\//, '').replace(/\//g, '-');
-
-      let additionalArgs = [];
-      if (options.posterPosition) {
-        additionalArgs = [...additionalArgs, '-ss', options.posterPosition];
-      }
-
-      if (options.posterScale) {
-        additionalArgs = [...additionalArgs, '-vf', `scale=${options.posterScale}`];
-      }
-
-      if (!ffmpeg.isLoaded()) {
-        await ffmpeg.load();
-      }
-
-      // Write file to MEMFS first so that ffmpeg.wasm is able to consume it
-      // eslint-disable-next-line new-cap
-      ffmpeg.FS('writeFile', ffmpegSrc, await fetchFile(srcPath));
-      await ffmpeg.run(
-        '-i',
-        ffmpegSrc,
-        ...additionalArgs,
-        '-vframes',
-        '1',
-        '-f',
-        'image2',
-        ffmpegDest
-      );
-      await promises.writeFile(
-        cachedDestPath,
-        // eslint-disable-next-line new-cap
-        ffmpeg.FS('readFile', ffmpegDest)
-      );
-    }
-
-    if (!timestampDest || timestampDest < timestampDestCached || timestampDest < timestampSrc) {
-      await promises.copyFile(cachedDestPath, destPath);
-    }
-
-    return dest;
-  };
-
-  const generatePosterImages = async () => {
-    const files = [...posterQueue];
-
-    const bar = new SingleBar(
-      {
-        format:
-          '    ➞ Generating poster images: [{bar}] {percentage}% | ETA: {eta}s | {value}/{total}',
-      },
-      Presets.legacy
-    );
-    let progress = 0;
-    // Start the progress bar with a total value of 200 and start value of 0
-    bar.start(files.length, 0);
-
-    // Ffmpeg wasm can't process more than one file at a time
-    // so we need to run it serial
-    for (const src of files) {
-      try {
-        const dest = getPosterImageFilePath(src);
-        // eslint-disable-next-line no-await-in-loop
-        await generatePosterImage(src, dest);
-        progress++;
-        bar.update(progress);
-      } catch (error) {
-        console.log(error);
-      }
-    }
-
-    bar.stop();
-    return true;
-  };
-
-  const getLocalPath = (src, sys, addToQueue = true) => {
-    const url = new URL(src);
-    const { updatedAt } = sys || {};
-    const { searchParams, pathname } = url;
-    const extname = pathExtname(pathname);
-    const filename = basename(pathname, extname);
-    const params = new URLSearchParams(searchParams);
-    const { w, h, f, fm = extname.substring(1) } = Object.fromEntries(params.entries());
-
-    const [, , assetId] = dirname(pathname).split('/');
-
-    const file = [filename, w ? `-w${w}` : '', h ? `-h${h}` : '', f ? `-${f}` : '', '.', fm]
-      .filter((v) => v)
-      .join('');
-
-    const queueEntry = {
-      src,
-      timestamp: updatedAt ? Date.parse(updatedAt) : 0,
-    };
-    if (addToQueue) {
-      queue.add(JSON.stringify(queueEntry));
-    }
-
-    return join(options.assetBase, assetId, file);
-  };
-
-  const getLocalSrc = (src, sys, addToQueue = true) => {
-    const localPath = getLocalPath(src, sys, addToQueue);
-    if (process.env.HUGO_BASEURL) {
-      return join('/', process.env.HUGO_BASEURL, localPath);
-    }
-
-    return localPath;
-  };
-
-  const readAsset = async (asset) => {
-    const { sys, fields } = asset || {};
-    const fileUrl = fields?.file?.url ?? '';
-    const src = fileUrl.startsWith('//') ? `https:${fileUrl}` : fileUrl;
-
-    if (!src) {
-      return;
-    }
-
-    const url = new URL(src);
-    const filepath = join(options.cachePath, getLocalPath(src, {}, false));
-    await mkdirp(dirname(filepath));
-
-    const assetTimestamp = await getAssetTimestamp(sys);
-    const fileTimestamp = await getFileTimestamp(filepath);
-
-    if (!existsSync(filepath) || !assetTimestamp || assetTimestamp > fileTimestamp) {
-      try {
-        const response = got(url);
-        const buffer = await response.buffer();
-
-        await promises.writeFile(filepath, buffer);
-      } catch {
-        console.log('Error downloading image:', url);
-      }
-    }
-
-    return promises.readFile(filepath, 'utf8');
-  };
-
-  const fetchAsset = async (src, timestamp) => {
-    const url = new URL(src);
-
-    const filepath = getLocalPath(src, {}, false);
-    const cacheFile = join(options.cachePath, filepath);
-    const file = join(options.assetPath, filepath);
-
-    await mkdirp(dirname(cacheFile));
-    await mkdirp(dirname(file));
-
-    const cachedFileTimestamp = await getFileTimestamp(cacheFile);
-    const fileTimestamp = await getFileTimestamp(file);
-
-    if (!existsSync(cacheFile) || (timestamp && timestamp > cachedFileTimestamp)) {
-      try {
-        const response = got(url);
-        const buffer = await response.buffer();
-        await promises.writeFile(cacheFile, buffer);
-      } catch {
-        console.log('Error downloading image:', url);
-      }
-    }
-
-    if (!fileTimestamp || fileTimestamp < cachedFileTimestamp || fileTimestamp < timestamp) {
-      await promises.copyFile(cacheFile, file);
-    }
-
-    return filepath;
-  };
-
-  const fetchAssets = async () => {
-    const files = [...queue].map((json) => JSON.parse(json));
-
-    const bar = new SingleBar(
-      { format: '    ➞ Fetching files: [{bar}] {percentage}% | ETA: {eta}s | {value}/{total}' },
-      Presets.legacy
-    );
-    let progress = 0;
-    // Start the progress bar with a total value of 200 and start value of 0
-    bar.start(files.length, 0);
-    await Promise.all(
-      files.map(async (file) => {
-        const { src, timestamp } = file;
-        try {
-          await fetchAsset(src, timestamp);
-          progress++;
-          bar.update(progress);
-        } catch (error) {
-          console.log(error);
-        }
-      })
-    );
-
-    bar.stop();
-    return true;
-  };
-
-  const getImageData = (asset: Asset, ratio, focusArea) => {
-    const { width, height } = asset?.fields?.file?.details?.image ?? {};
-    const mimeType = asset?.fields?.file?.contentType;
-    const url = asset?.fields?.file?.url;
-    const types = [...new Set([...(options.extraTypes || []), mimeType])];
-    const sizes = (options.sizes || []).map((value) => {
-      if (typeof value === 'function') {
-        return value(asset, ratio, focusArea);
-      }
-
-      return parseFloat(value.toString());
-    });
-
-    if (!url || !width || !height) {
-      return {};
-    }
-
-    const src = url.startsWith('//') ? `https:${url}` : url;
-
-    // Compute possible widths considering ratio & contentful max size
-    const widths = [
-      ...new Set(
-        (ratio
-          ? [
-              contentfulMaxWidth,
-              contentfulMaxWidth * ratio,
-              width,
-              Math.floor(height * ratio),
-              ...sizes,
-            ]
-          : [contentfulMaxWidth, width, ...sizes]
-        )
-          .sort((a, b) => b - a)
-          .filter(
-            (w) =>
-              w <= width &&
-              w <= contentfulMaxWidth &&
-              (!ratio || (w / ratio <= contentfulMaxWidth && w / ratio <= height))
-          )
-          .map((w) => Math.round(w))
-      ),
-    ];
-
-    const [maxWidth] = widths;
-    const maxHeight = Math.round(maxWidth / (ratio || width / height));
-    const sizesAttribute = `(max-width: ${maxWidth}px) 100vw, ${maxWidth}px`;
-
-    const sizeParams = (width, ratio) => {
-      if (!ratio) {
-        return `w=${width}`;
-      }
-
-      return [
-        'fit=fill',
-        `w=${width}`,
-        `h=${Math.floor(width / ratio)}`,
-        `f=${focusArea || 'center'}`,
-      ].join('&');
-    };
-
-    const megaPixelFilter = (w, type = '-') => {
-      const max = maxMegaPixels?.[type.replace('image/', '')];
-      const r = ratio || width / height;
-      return !max || max >= (w * Math.floor(w / r)) / 1000000;
-    };
-
-    const sources = Object.fromEntries(
-      types.map((type) => {
-        const fm = type === mimeType ? '' : `&fm=${type.replace('image/', '')}`;
-        return [
-          type,
-          widths
-            .filter((w) => megaPixelFilter(w, type))
-            .map((w) => ({
-              src: `${src}?${sizeParams(w, ratio)}&q=${options.quality}${fm}`,
-              width: w,
-            })),
-        ];
-      })
-    );
-
-    const [assetFile] = sources[mimeType];
-    const { src: assetSrc } = assetFile;
-
-    // Fetch original src and srcsets for production use
-    const finalSrc = options.download ? getLocalSrc(assetSrc, asset.sys) : assetSrc;
-    const finalSrcsets = Object.entries(sources).map(([type, files]) => {
-      const srcset = files.map((file) => {
-        const { width, src } = file;
-        return `${options.download ? getLocalSrc(src, asset.sys) : src} ${width}w`;
-      });
-      return {
-        type,
-        srcset: srcset.join(', '),
-      };
-    });
-
-    return {
-      width: maxWidth,
-      height: maxHeight,
-      sizes: sizesAttribute,
-      srcsets: finalSrcsets,
-      src: finalSrc,
-    };
-  };
+  const assetHelper = getAssetHelper(options);
+  const svgHelper = getSvgHelper(options);
+  const imageHelper = getImageHelper(options);
+  const videoHelper = getVideoHelper(options);
 
   const mapAssetLink = async (
     transformContext: TransformContext,
     _runtimeContext,
     defaultValue: MapAssetLink
   ): Promise<ProcessedAsset | ProcessedSvg | ProcessedImage | ProcessedVideo> => {
-    const { download } = options || {};
-    const { asset, entry, fieldId } = transformContext;
+    const { download } = options;
+    const { asset } = transformContext;
     const { sys } = asset || {};
     const { url, mimeType = '' } = defaultValue;
     const src = url.startsWith('//') ? `https:${url}` : url;
 
-    // Get ratio from config
-    const defaultRatio = entry?.fields?.ratio ?? options?.ratios?.default;
-    const contentTypeDefaultRatio =
-      options?.ratios?.[entry?.sys?.contentType?.sys?.id ?? 'unknown']?.default ?? defaultRatio;
-
-    const { [entry?.sys?.contentType?.sys?.id ?? 'unknown']: contentTypeRatios } =
-      options?.ratios?.contentTypes ?? {};
-    const ratioConfig = contentTypeRatios?.fields?.[fieldId] ?? contentTypeDefaultRatio;
-
-    // Get focusArea from config
-    const defaultFocusArea = entry?.fields?.focus_area ?? options?.focusAreas?.default ?? 'center';
-    const contentTypeDefaultFocusArea =
-      options?.focusAreas?.[entry?.sys?.contentType?.sys?.id ?? 'unknown']?.default ??
-      defaultFocusArea;
-    const { [entry?.sys?.contentType?.sys?.id ?? 'unknown']: focusAreaConfig } =
-      options?.focusAreas?.contentTypes ?? {};
-    const focusArea = focusAreaConfig?.fields?.[fieldId] ?? contentTypeDefaultFocusArea;
-
     if (mimeType === 'image/svg+xml') {
-      const source = await readAsset(asset);
-
-      const { data: optimized } = await optimize(source, {
-        multipass: true,
-        plugins: [
-          {
-            name: 'preset-default',
-            params: {
-              overrides: {
-                removeViewBox: false,
-              },
-            },
-          },
-          'removeDimensions',
-          {
-            name: 'removeAttrs',
-            params: {
-              attrs: 'fill',
-            },
-          },
-        ],
-      });
-
-      // Add viewBox if not present
-      return {
-        ...defaultValue,
-        source: optimized,
-        srcsets: [],
-        src: download ? getLocalSrc(src, sys) : src,
-      } as ProcessedSvg;
+      return svgHelper.mapAssetLink(transformContext, defaultValue);
     }
 
     if (mimeType.startsWith('image')) {
-      const original = getImageData(asset, undefined, focusArea);
-      const derivatives = Object.fromEntries(
-        Object.entries(ratioConfig || {}).map(([name, ratio]) => [
-          name,
-          getImageData(asset, ratio, focusArea),
-        ])
-      );
-      return {
-        ...defaultValue,
-        src: original.src,
-        derivatives: { original, ...derivatives },
-      } as ProcessedImage;
+      return imageHelper.mapAssetLink(transformContext, defaultValue);
     }
 
-    if (mimeType.startsWith('video') && options.generatePosterImages) {
-      const filepath = getLocalPath(src, sys, !options.download);
-
-      posterQueue.add(filepath);
-
-      return {
-        ...defaultValue,
-        src: download ? getLocalSrc(src, sys) : src,
-        poster: getPosterImageFilePath(filepath),
-      } as ProcessedVideo;
+    if (mimeType.startsWith('video')) {
+      return videoHelper.mapAssetLink(transformContext, defaultValue);
     }
 
-    return { ...defaultValue, src: download ? getLocalSrc(src, sys) : src } as ProcessedAsset;
+    return {
+      ...defaultValue,
+      src: download ? assetHelper.getLocalSrc(src, sys) : src,
+    } as ProcessedAsset;
   };
 
   const after = async () => {
-    if (options.download || options.generatePosterImages) {
-      await fetchAssets();
-    }
-
-    if (options.generatePosterImages) {
-      await generatePosterImages();
-    }
+    await assetHelper.after();
+    await imageHelper.after();
+    await svgHelper.after();
+    await videoHelper.after();
   };
 
   return {
